Finish wallet address in download
