+++ packages/react-app/tests/integration/api/bounty.claim.spec.ts
@@ -2,7 +2,8 @@ import { NextApiRequest, NextApiResponse } from 'next';
 import { createMocks, MockResponse, MockRequest } from 'node-mocks-http';
 import { Connection } from 'mongoose';
 import dbConnect from '@app/utils/dbConnect';
-import Bounty, { ActivityHistoryItem, BountyBoardSchema, StatusHistoryItem } from '@app/models/Bounty';
+import { ActivityHistoryItem, StatusHistoryItem } from '@app/models/Bounty';
+import Bounty, { BountyBoardSchema } from '@app/models/BountyDb';
 import bountyClaimHandler from '@app/pages/api/bounties/[id]/claim';
 import { testBounty } from '@tests/stubs/bounty.stub';
 import { newActivityHistory, newStatusHistory } from '@app/utils/formUtils';
