+++ packages/react-app/tests/integration/api/bounty.paid.spec.ts
@@ -2,7 +2,8 @@ import { NextApiRequest, NextApiResponse } from 'next';
 import { createMocks, MockResponse, MockRequest } from 'node-mocks-http';
 import { Connection } from 'mongoose';
 import dbConnect from '@app/utils/dbConnect';
-import Bounty, { ActivityHistoryItem, BountyBoardSchema } from '@app/models/Bounty';
+import { ActivityHistoryItem } from '@app/models/Bounty';
+import Bounty, { BountyBoardSchema } from '@app/models/BountyDb';
 import bountyPaidHandler from '@app/pages/api/bounties/[id]/paid';
 import { testBounty } from '@tests/stubs/bounty.stub';
 import { newActivityHistory } from '@app/utils/formUtils';
