+++ packages/react-app/tests/integration/api/bounty.spec.ts
@@ -2,7 +2,8 @@ import { NextApiRequest, NextApiResponse } from 'next';
 import { createMocks, MockResponse, MockRequest } from 'node-mocks-http';
 import { Connection } from 'mongoose';
 import dbConnect from '@app/utils/dbConnect';
-import Bounty, { BountyCollection, BountyBoardSchema } from '@app/models/Bounty';
+import { BountyCollection } from '@app/models/Bounty';
+import Bounty, { BountyBoardSchema } from '@app/models/BountyDb';
 import bountiesHandler from '@app/pages/api/bounties';
 import bountyHandler from '@app/pages/api/bounties/[id]';
 import { testBounty } from '@tests/stubs/bounty.stub';
